translated my optimization from mathematica to python
very slow, try pdv = (5,3,8). this takes about 30 minutes. mathematica takes 10 seconds for the same task. would be great if we can improve this.
as we said a big issue is the processing/substituting in sympy. maybe we can do as little as possible of this (we might have to change sympyMinimizer for this).